// ui/src/generated/NoteEdit.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Beats } from "./Beats";
import type { NoteId } from "./NoteId";

export type NoteEdit = { id: NoteId, pitch: number | null, velocity: number | null, start: Beats | null, duration: Beats | null, muted: boolean | null, };